requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow==2.19.0
pillow==10.1.0
numpy==1.26.2
python-multipart==0.0.6
jinja2==3.1.2
requests==2.31.0
